fix default meta translations
