+++ documentation-website/src/components/t.ts
@@ -1,0 +1,28 @@
+import languageKey from '../locales/language-key.ts';
+import Window from './window.ts';
+import {
+  FIRST_ELEMENT,
+  SECOND_ELEMENT,
+} from '../constants.ts';
+import files from '../locales/files.ts';
+
+// eslint-disable-next-line complexity
+export default async(lnkey: languageKey, global?: object,): Promise<string> => {
+  const that: Window = (global || window) as Window;
+  const language = (that?.Navigator?.language ?? 'en')
+    .replace(/-.*$/u, '',);
+  const main = lnkey.split('.',)[FIRST_ELEMENT];
+  if (!files.includes(`en-${main}`,)) {
+    return lnkey;
+  }
+  const originals = await import(`../locales/en-${ main }.ts`);
+  let output = (files.includes(`${ language }-${ main }`,)
+    ? await import(`../locales/${ language }-${ main }.ts`)
+    : originals).default;
+  let defaultOutput = originals.default;
+  for (const part of lnkey.split('.',).slice(SECOND_ELEMENT,)) {
+    output = output[part] || defaultOutput[part];
+    defaultOutput = defaultOutput[part];
+  }
+  return output || lnkey;
+};
